requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.25.2
pypdf==5.8.0
pydantic==2.10.3
python-multipart==0.0.6
python-dotenv==1.0.0



pip install langchain langchain-experimental langchain-community sentence-transformers